requirements.txt
discodo
discord.py
python-dotenv
jishaku